General: Future proof and disregard string return type. It never happened before

+++ Day1.h
@@ -6,7 +6,7 @@
 
 namespace aoc {
 namespace Day1 {
-int Part1(const std::vector<int>& data) {
+uint64_t Part1(const std::vector<int>& data) {
 	int last{data[0]};
 	int sum = std::count_if(data.begin(), data.end(), [&](const int depth){
 		bool res = depth > last;
@@ -16,7 +16,7 @@ int Part1(const std::vector<int>& data) {
 	return sum;
 }
 
-int Part2(const std::vector<int>& data) {
+uint64_t Part2(const std::vector<int>& data) {
 	int sum{0};
 	for (size_t i{0}; i < data.size(); i++) {
 		if (i + 3 < data.size() && data[i + 3] > data[i]) {
